% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_us_states.R
\docType{data}
\name{covid_us_states}
\alias{covid_us_states}
\title{Provides the Dataset for the Confirmed Covid-19 Cases and Deaths by States in the U.S, reported by The New York Times.}
\format{
a tibble with 11,904 observations and 5 variables:
\itemize{
\item \strong{date}: Date reported
\item \strong{state}: Name of U.S. state
\item \strong{fips}: Two-letter state code
\item \strong{cases}: Cumulative number of cases per state
\item \strong{deaths}: Cumulative number of deaths per state
}
}
\source{
\href{https://github.com/nytimes/covid-19-data}{NYtimes GitHub}
}
\usage{
covid_us_states
}
\description{
Provides the Dataset for the Confirmed Covid-19 Cases and Deaths by States in the U.S, reported by The New York Times.
}
\keyword{datasets}
