% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datatable.R
\name{create_datatable}
\alias{create_datatable}
\title{Creates a list of 2 elements relating to the spatial data of Covid-19 testing in the U.S}
\usage{
create_datatable()
}
\description{
Creates a list of 2 elements relating to the spatial data of Covid-19 testing in the U.S
}
