% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_data.R
\name{create_map_data}
\alias{create_map_data}
\title{Creates a list of 2 elements relating to the spatial data of Covid-19 cases in the U.S}
\usage{
create_map_data()
}
\description{
Creates a list of 2 elements relating to the spatial data of Covid-19 cases in the U.S
}
