% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custome_theme}
\alias{custome_theme}
\title{Custom theme for maps and bar charts}
\usage{
custome_theme(plot_type)
}
\arguments{
\item{plot_type}{The type of plot - "map" or "bar"}
}
\value{
Custom theme that complements the aesthetics of the Shiny App
}
\description{
Custom theme for maps and bar charts
}
