% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_msg.R
\name{error_msg}
\alias{error_msg}
\title{Creates error message if Shiny App input is empty}
\usage{
error_msg(input, plot_type)
}
\arguments{
\item{input}{Shiny user-interface input}

\item{plot_type}{"bar" or "map"}
}
\value{
Error message if condition(s) are not met in the Shiny App
}
\description{
Creates error message if Shiny App input is empty
}
