% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_raw.R
\docType{data}
\name{tests_raw}
\alias{tests_raw}
\title{Reported Covid-19 Testing Data across all U.S States from 2020-01-22 to 2020-10-07}
\format{
a tibble with 11,164 observations and 5 variables:
\itemize{
\item \strong{date}: Date reported
\item \strong{state}: Abbreviated name of U.S. state
\item \strong{daily_cases}: daily number of Covid-19 cases per U.S state
\item \strong{daily_tests}: Daily number of Covid-19 testing per U.S state
\item \strong{total_cases}: Total number of cases per U.S State
\item \strong{total_tests}: Total number of Covid-19 testing per U.S state
\item \strong{data_quality}: Quality of Covid-19 testing data provided by the U.S state
}
}
\source{
\href{https://github.com/nytimes/covid-19-data}{NYtimes GitHub}
}
\usage{
tests_raw
}
\description{
Reported Covid-19 Testing Data across all U.S States from 2020-01-22 to 2020-10-07
}
\keyword{datasets}
