% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_pop_ests.R
\docType{data}
\name{us_pop_ests}
\alias{us_pop_ests}
\title{Provides the dataset for the estimated population by U.S county, primarily used for relational joins.}
\format{
a tibble with 51 observations and 2 variables
\itemize{
\item \strong{fips}: Two-letter U.S state code
\item \strong{population}: Total population estimates
}
}
\source{
\href{https://github.com/kingaa/covid-19-data}{Covid-19 data on Github}
}
\usage{
us_pop_ests
}
\description{
Provides the dataset for the estimated population by U.S county, primarily used for relational joins.
}
\keyword{datasets}
