% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_states_map.R
\docType{data}
\name{us_states_map}
\alias{us_states_map}
\title{Raw US map data for states}
\format{
a tibble with 12,999 observations and 5 variables
\itemize{
\item \strong{lat}: Latitude of the U.S state
\item \strong{longitude}: Longitude of the U.S state
\item \strong{group}: Geographical coordinates for a group of observations/states
\item \strong{fips}: Two-letter U.S state code
\item \strong{abbr}: Abbreviated U.S state name
}
}
\source{
usmap R package
}
\usage{
us_states_map
}
\description{
Raw US map data for states
}
\keyword{datasets}
