% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_data.R
\name{create_map_data}
\alias{create_map_data}
\title{Uses relational joins to create spatial data for Covid-19 cases and tests data embedded in the App}
\usage{
create_map_data()
}
\description{
Uses relational joins to create spatial data for Covid-19 cases and tests data embedded in the App
}
