% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_msg.R
\name{error_msg}
\alias{error_msg}
\title{Title}
\usage{
error_msg(input, plot_type)
}
\arguments{
\item{input}{Shiny user-interface input}
}
\value{
Error message if condition(s) are not met
}
\description{
Title
}
\examples{
error_msg(input$choose_states, plot_type == "map")
}
