% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_input.R
\name{ui_input}
\alias{ui_input}
\title{Creates reactive input for Shiny app}
\usage{
ui_input(input_type, input_id, data)
}
\arguments{
\item{input_type}{Character string specifying type of input, "date" or "state"}

\item{input_id}{A string specifying the label Id}

\item{data}{A data frame}
}
\value{
Creates a reactive slider/picker input for Shiny App
}
\description{
Creates reactive input for Shiny app
}
